Update day 9-1.cpp

+++ adventofcode/day 9-1.cpp
@@ -20,7 +20,7 @@ int n=s.length();
         }
     }
     reverse(avail.begin(),avail.end());
-    for(int i=(int) v.size()-1;i>=0;i--){
+    for(int i= v.size()-1;i>=0;i--){
         if(v[i]!=0){
             if(!avail.empty() && avail.back()<i){
                 v[avail.back()]=v[i];
